AudioHandler beta :), see details
Basically it has a matrix of sounds, channels and paths, that loads/unlo
ads automatically many sound files with a fixed filename syntax. Sounds
are grouped in classifications that will help you to use  sounds in a ea
sy  and comprehensive way. In order to play/stop/pause/resume a sound, o
nly index and classification is needed. It comes with a set of template
sounds just to show you how it works. See details in the header of Audio
Handler class. There is a simple implementation in main.cpp